requirements.txt
gql==3.5.0
aiohttp==3.9.3
requests~=2.31.0
Flask==3.0.1
flask-cors
peewee~=3.17.0
gunicorn
werkzeug~=3.0.0
Flask-Bcrypt
PyJWT~=2.8.0
APScheduler~=3.10.4
PyYAML~=6.0.1